Added popover to bottom menu buttons
