+++ src/app/(frontend)/articles/[slug]/page.tsx
@@ -10,7 +10,7 @@ import { LivePreviewListener } from '@/components/LivePreviewListener'
 import RelatedArticles from '@/components/Article/RelatedArticles'
 import TopMenuContainer from '@/components/Article/TopMenu/TopMenuContainer'
 import RightMenuContainer from '@/components/Article/RightMenuContainer'
-import BottomMenu from '@/components/Article/BottomMenu'
+import BottomMenu from '@/components/BottomMenu'
 import { ArticleHero } from '@/heros/ArticleHero'
 import { RenderArticleBlocks } from '@/blocks/RenderArticleBlocks'
 import LeftMenuContainer from '@/components/Article/LeftMenu/LeftMenuContainer'
