+++ src/app/(frontend)/page.tsx
@@ -6,7 +6,7 @@ import PageClient from './page.client'
 import { LivePreviewListener } from '@/components/LivePreviewListener'
 import NotFound from './not-found'
 import RightMenuContainer from '@/components/Article/RightMenuContainer'
-import BottomMenu from '@/components/Article/BottomMenu'
+import BottomMenu from '@/components/BottomMenu'
 import { RenderHomeBlocks } from '@/blocks/RenderHomeBlocks'
 import { Home as HomeCollectionType } from '@/payload-types'
 import { getCachedGlobal } from '@/utilities/getGlobals'
